+++ docs/changelog/changelog.md
@@ -5,7 +5,7 @@
 ### Features
 
 - A new example has been created that is much easier to follow and show cases the features of dandy.
-- Added a new `Intel` class called `BaseIterableIntel` that is used to create an iterable intel object.
+- Added a new `Intel` class called `BaseListIntel` that is used to create an iterable intel object that behaves like a `list`.
 
 ### Changes